requirements.txt
pandas>=1.3.0
numpy>=1.20.0
pymysql>=1.0.2
cryptography>=3.4.0
websocket-client>=1.2.1
requests>=2.26.0
scikit-learn>=1.0.0
matplotlib>=3.4.0
seaborn>=0.11.0
PyQt5>=5.15.0 # 可选，用于GUI开发
tensorflow>=2.8.0 # 可选，用于AI模型
torch>=1.10.0 # 可选，用于AI模型 